Creating a build-log/change-log I
The output of the sync process is now written to a file in ~/android/
and we can extract the changes via, e.g., cat sync-17.1.log | grep -A 1 "Von".
Currently the log is appended to.
TODO: Generate a proper list of all changes and delete the sync-log
after a build is complete.