Creating a build-log/change-log III
After a successful build, the changes/repopicks are written to a file in
the same directory than where the final output is moved to.
Currently the change-log may contain the same entry multiple times due
to multiple repo sync runs prior to a build.
The sync-log is deleted.
TODO: Clean up the change-log. Remove multi-entries and eventually
generate proper urls that point to the actual changes.